// src/Backoffice/mipg/dto/direct-mipg.dto.ts
export class MipgDirectApiDto {
  id: string;
  mipg: string;
  psp: number;
  loginaccount: string;
  customercode: string;
  merchantconfigid: string;
  iv: string;
  key: string;
  terminal: string;
  merchant: string;
  status: any;
  karmozd: number;
  type: number;
  acceptorcode: string;
  sheba: string;
}